% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregation.R
\name{Aggregation}
\alias{Aggregation}
\title{Aggregation parameter for negative binomial distribution}
\usage{
Aggregation(A1, A2, Z, hybridIndex)
}
\arguments{
\item{A1}{A number. Aggregation in one side of hybrid index}

\item{A2}{A number. Aggregation in the other side of hybrid index}

\item{Z}{A number}

\item{hybridIndex}{A number}
}
\value{
The aggregation parameter of negative binomial at one given point of \code{hybridIndex}
}
\description{
Aggregation parameter for negative binomial distribution
}
