% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bananaPlots_alpha.R
\name{bananaPlots_alpha}
\alias{bananaPlots_alpha}
\title{Plot function}
\usage{
bananaPlots_alpha(mod, data, response, hybridIndex = seq(0, 1, 0.01),
  cols = c("green", "orange"), group, islog10 = F)
}
\arguments{
\item{mod}{A model to be ploted}

\item{data}{A data frame}

\item{response}{A character string. Response (e.g. "worm_count")}

\item{hybridIndex}{A vector of points representing the index used as x axis}

\item{cols}{A vector of two colors for the two groups}

\item{group}{A character string. Which group is considered (e.g. "Sex")}

\item{islog10}{logical. Is the y axis ploted with log10 transformation}
}
\value{
A plot of our model
}
\description{
Plot function
}
