% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitWeibull.R
\name{FitBasicNoAlphaWeibull}
\alias{FitBasicNoAlphaWeibull}
\title{Fit the model for Weibull distribution of the expected load}
\usage{
FitBasicNoAlphaWeibull(data, response, hybridIndex, paramBounds, config)
}
\arguments{
\item{data}{A data frame}

\item{response}{A character string. Response (e.g. "worm_count")}

\item{hybridIndex}{A vector of points representing the index used as x axis}

\item{paramBounds}{A vector of parameters (upper, lower, start) for the optimisation}

\item{config}{A list containing an optimizer (default: "optimx"), a method (default "bobyqa", "L-BFGS-B") and a control (default list(follow.on = TRUE))}
}
\value{
A fit for binomial distributed data
}
\description{
Fit the model for Weibull distribution of the expected load
}
