% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBananaDF.R
\name{getBananaDF}
\alias{getBananaDF}
\title{Get bananaDF for a given model}
\usage{
getBananaDF(mod, hybridIndex)
}
\arguments{
\item{mod}{A model to be ploted}

\item{hybridIndex}{A vector of points representing the index used as x axis}
}
\value{
A data frame to be ploted
}
\description{
Get bananaDF for a given model
}
