% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamBounds.R
\name{getParamBounds}
\alias{getParamBounds}
\title{Get our parameters at strat, lower, upper, for optimisation}
\usage{
getParamBounds(model, data, response)
}
\arguments{
\item{model}{Method to be used in fitting the model
So far implemented for "binomial", "negbin", "student"}

\item{data}{A data frame}

\item{response}{A character string. Response (e.g. "worm_count")}
}
\value{
A vector of parameters (upper, lower, start) for the optimisation
}
\description{
Get our parameters at strat, lower, upper, for optimisation
}
