% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserInput.R
\name{glm.hybrid}
\alias{glm.hybrid}
\title{glm.hybrid}
\usage{
glm.hybrid(formula, data, alpha.along, alpha.start = 0.1,
  start.mod = MASS::glm.nb, start.values = NA)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{alpha.along}{}

\item{alpha.start}{}

\item{start.mod}{}

\item{start.values}{}
}
\value{
an object of class glm.hybrid
}
\description{
Here we describe our main formula interface
}
\details{
And here we put all the details
}
\author{
Emanuel Heitlinger
}

