% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshallData.R
\name{marshallData}
\alias{marshallData}
\title{Prepare our data for fitting by group}
\usage{
marshallData(data, response, group)
}
\arguments{
\item{data}{A data frame}

\item{response}{A character string. Response (e.g. "worm_count")}

\item{group}{A character string. Which group is considered (e.g. "Sex")}
}
\value{
A list containing the full dataset, a dataset for groupA, a dataset for groupB
}
\description{
Prepare our data for fitting by group
}
