% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanLoad.R
\name{MeanLoad}
\alias{MeanLoad}
\title{Function defining the expected load at one given point of the index}
\usage{
MeanLoad(L1, L2, alpha, hybridIndex)
}
\arguments{
\item{L1}{A number. Aggregation in one side of hybrid index}

\item{L2}{A number. Aggregation in the other side of hybrid index}

\item{alpha}{A number. Curvature of the expected load along hybrid index}

\item{hybridIndex}{A number}
}
\value{
The expected load at one given point of \code{hybridIndex}
}
\description{
Function defining the expected load at one given point of the index
}
