% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlHyb.R
\name{mlHyb}
\alias{mlHyb}
\title{Run analysis in user friendly formula interface}
\usage{
mlHyb(formula, data, model, hybridIndex = "HI",
  myparamBounds = "default", hybridEffect = TRUE,
  config = list(optimizer = "optimx", method = c("L-BFGS-B", "bobyqa"),
  control = list(follow.on = TRUE)))
}
\arguments{
\item{formula}{An object of class "formula"}

\item{data}{A data frame containing the variables in the model}

\item{model}{Method to be used in fitting the model
So far implemented for "binomial", "negbin", "student"}

\item{hybridIndex}{The hybrid index}

\item{myparamBounds}{A named vector giving the start, lower and upper parameters for optimization}

\item{hybridEffect}{Logical. TRUE if hybrid effect (deviation from additivity between both parents) is tested, FALSE if not.}

\item{group}{A character string. Which group is considered (e.g. "Sex")}
}
\value{
mlHyb returns an object of class "mlHyb"
The function print creates a user friendly output
The functions summary is usedto obtain a likelihood ratio test (G-test) between models with and without bend (hybridization effect)
An object of class "mlHyb" is a list containing at least the following components:
  fit : the ML fit considering hybridization effect
  fitNull : the ML fit considering perfect additivity between both sides (null hypothesis on hybridization effect)
  call : the call
}
\description{
Run analysis in user friendly formula interface
}
