% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printConvergence.R
\name{printConvergence}
\alias{printConvergence}
\title{Print if model converged or not}
\usage{
printConvergence(fit)
}
\arguments{
\item{fit}{A fit to be tested for convergence}
}
\value{
A character string, "Did converge" or "Did not converge"
}
\description{
Print if model converged or not
}
