% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAll.R
\name{runAll}
\alias{runAll}
\title{Run full analysis for the 4 hypotheses}
\usage{
runAll(data, response, model, group, paramBounds)
}
\arguments{
\item{data}{A data frame}

\item{response}{A character string. Response (e.g. "worm_count")}

\item{model}{Method to be used in fitting the model
So far implemented for "binomial", "negbin", "student"}

\item{group}{A character string. Which group is considered (e.g. "Sex")}
}
\value{
The full result of the fit
}
\description{
Run full analysis for the 4 hypotheses
}
