% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWeibullShifted.R
\name{runWeibullShifted}
\alias{runWeibullShifted}
\title{Fit the four cases for weibull distribution of the expected load}
\usage{
runWeibullShifted(data, response, hybridIndex, paramBounds, config)
}
\arguments{
\item{data}{A data frame}

\item{response}{A character string. Response (e.g. "worm_count")}

\item{hybridIndex}{A vector of points representing the index used as x axis}

\item{paramBounds}{A vector of parameters (upper, lower, start) for the optimisation}

\item{config}{A list containing an optimizer (default: "optimx"), a method (default "bobyqa", "L-BFGS-B") and a control (default list(follow.on = TRUE))}
}
\value{
A fit for binomial distributed data for all four cases
}
\description{
Four hypotheses tested : no difference between subspecies, no hybrid effect;
no difference between subspecies, flexible hybrid effect
difference between subspecies, no hybrid effect
difference between subspecies, flexible hybrid effect
}
