% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SizeNegBin.R
\name{SizeNegBin}
\alias{SizeNegBin}
\title{Size parameter for negative binomial distribution}
\usage{
SizeNegBin(A1, A2, Z, hybridIndex)
}
\arguments{
\item{A1}{A number. Aggregation in one side of hybrid index}

\item{A2}{A number. Aggregation in the other side of hybrid index}

\item{Z}{A number}

\item{hybridIndex}{A number}
}
\value{
The size parameter of negative binomial at one given point of \code{hybridIndex}
}
\description{
Size parameter for negative binomial distribution
}
