% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlHyb.R
\name{mlHyb}
\alias{mlHyb}
\title{Run analysis in user friendly formula interface}
\usage{
mlHyb(formula, data, model, hybridIndex = "HI",
  myparamBounds = "default", hybridEffect = TRUE,
  config = list(optimizer = "optimx", method = c("L-BFGS-B", "bobyqa"),
  control = list(follow.on = TRUE)))
}
\arguments{
\item{formula}{An object of class "formula"}

\item{data}{A data frame containing the variables in the model}

\item{model}{Method to be used in fitting the model
So far implemented for "binomial", "negbin", "student"}

\item{hybridIndex}{The hybrid index}

\item{myparamBounds}{A named vector giving the start, lower and upper parameters for optimization}

\item{hybridEffect}{Logical. TRUE if hybrid effect (deviation from additivity between both parents) is tested, FALSE if not.}

\item{group}{A character string. Which group is considered (e.g. "Sex")}
}
\value{
mlHyb returns an object of class "mlHyb"
The functions summary and anova are used to obtain and print a summary and analysis of variance table of the results. The generic accessor functions coefficients, effects, fitted.values and residuals extract various useful features of the value returned by lm.
An object of class "mlHyb" is a list containing at least the following components:
  coefficients
TO BE UPDATED
}
\description{
Run analysis in user friendly formula interface
}
